requirements.txt
Flask==3.1.1
gunicorn==21.2.0
flask-cors==6.0.1
opencv-python-headless==4.12.0.88
numpy==2.3.2
scikit-learn==1.7.1
Pillow==11.3.0
tensorflow-cpu==2.18.2
matplotlib==3.10.5
seaborn==0.13.2
pandas==2.3.1
python-dotenv==1.0.0
flask-limiter==3.8.1
